+++ sdk/computefleet/arm-computefleet/samples/v2-beta/javascript/operationsListSample.js
@@ -12,7 +12,7 @@ const { DefaultAzureCredential } = require("@azure/identity");
  */
 async function operationsList() {
   const credential = new DefaultAzureCredential();
-  const subscriptionId = "00000000-0000-0000-0000-00000000000";
+  const subscriptionId = "00000000-0000-0000-0000-000000000000";
   const client = new AzureFleetClient(credential, subscriptionId);
   const resArray = new Array();
   for await (const item of client.operations.list()) {